# swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is the API for the backend server (orchestrator) of the US Post prototype for the ACIT Cloud Studio
  version: 1.0.0
  title: Swagger USPS
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: thomas.i.jacobs@accenture.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: postal.mybluemix.net
basePath: /
...
schemes:
  - http
paths:
...
  /update-dispatch-settlement:
    post:
      tags:
        - Blockchain
      summary: Updates settlement status for packages
      description: >-
        This endpoint will take the dispatchId to update the settlement status. Fill in the type variable of the updateSettlement object as 'dispatch'
      operationId: updateSettlementStatus
      produces:
        - application/json
      parameters:
        - in: body
          name: updateSettlementObject
          description: parameters to update settlement status
          required: true
          schema:
            $ref: '#/definitions/UpdateDispatchSettlement'
      responses:
        '200':
          description: Updated Dispatch settlement type. Returning updated dispatch object
          schema: 
            type: array
            items:
                $ref: '#/definitions/Package'
        '405':
          description: Was not able to get simulated data
  /update-package-settlement:
    post:
      tags:
        - Blockchain
      summary: Updates settlement status for packages
      description: >-
        This endpoint will take the packageId to update the settlement status
      operationId: updateSettlementStatus
      produces:
        - application/json
      parameters:
        - in: body
          name: updateSettlementObject
          description: parameters to update settlement status. Fill in the type variable of the updateSettlement object as 'package'
          required: true
          schema:
            $ref: '#/definitions/UpdateSettlement'
      responses:
        '200':
          description: Updated package settlement status. Returning updated package object
          schema: 
            $ref: '#/definitions/Package'
        '405':
          description: Was not able to get simulated data
  /package-history:
    get:
      tags:
        - Blockchain
      summary: Gets a package's transaction history
      description: >-
        This endpoint will take a packageId and return the transaction history from the blockchain
      operationId: PackageHistory
      produces:
        - application/json
      parameters:
        - in: query
          name: packageId
          description: size of the simulation
          required: true
          type: string
      responses:
        '200':
          description: Got package history.
          schema: 
            type: array
            items:
                $ref: '#/definitions/PackageHistory'
        '405':
          description: Was not able to get package history data  
definitions:
  PackageHistory:
    type: object
    properties:
      date:
        type: string
      status:
        type: string
      statusType:
        type: string
        description: 'Either Settlement Status or Shipment Status'
      statusDescription:
        type: string
        required: false
      performedBy:
        type: string
        required: false
        description: 'Postal Org that initiated Settlement Status'
  Package:
    type: object
    properties:
      packageId:
        type: string
      weight:
        type: number
      type:
        type: string
      dispatchId:
        type: string
      receptacleId:
        type: string
      settlementStatus:
        type: string
      shipmentStatus:
        type: string
...
  UpdateSettlement:
    type: object
    properties:
      type:
        type: string
      id:
        type: string
      newStatus:
        type: string
      post:
        type: string 
      country:
        type: string
  UpdateDispatchSettlement:
    type: object
    properties:
      type:
        type: string
      startDate:
        type: date
        required: false
      endDate:
        type: date
        required: false
      originPost: 
        type: string
        required: false
      destinationPost:
        type: string
        required: false
      dateCreated:
        type: date
        required: false
      packageType:
        type: string
        required: false
      dispatchId:
        type: string
      newStatus:
        type: string
      post:
        type: string 
      country:
        type: string
...